add a RO websocket with no input handling
The goal here is to provide a mechanism to ensure a view-only
connection.

Currently, this is handled on the frontend, meaning as long as the
websocket is open, it is possible to send keystrokes through.

Adding another endpoint allows for clean separation and smooth
integration with reverse proxies/middlewares/auth setups out there.

This helps a use case where a user has got an established setup around
that or wants to integrate with their platform - so custom auth
middlewares can smoothly handle routing users between RW/RO endpoints.